+++ api/models/types.ts
@@ -1,4 +1,10 @@
+import { SEO } from 'api/types';
+
 export interface Model {
-	id: number;
-	name: string;
+    id: number;
+    name: string;
+    slug: string;
+    seoSpareParts: SEO;
+    seoWheels: SEO;
+    seoCabins: SEO;
 }
